refactor transaction manager service and store interface
