+++ store/leveldb/leveldb.go
@@ -7,15 +7,14 @@ import (
 
 	"github.com/brunotm/replicant/store"
 	"github.com/brunotm/replicant/transaction"
-	"github.com/brunotm/replicant/transaction/manager"
 	"github.com/syndtr/goleveldb/leveldb"
 )
 
-var _ manager.Store = (*Store)(nil)
+var _ store.Store = (*Store)(nil)
 
 func init() {
 	store.Register("leveldb",
-		func(uri string) (s manager.Store, err error) {
+		func(uri string) (s store.Store, err error) {
 			return New(uri)
 		})
 }
