+++ store/store.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/brunotm/replicant/transaction/manager"
+	"github.com/brunotm/replicant/transaction"
 )
 
 var (
@@ -15,8 +15,31 @@ var (
 	ErrTransactionNotFound = errors.New("transaction not found")
 )
 
+// Store for transaction configurations
+type Store interface {
+
+	// Close the store
+	Close() (err error)
+
+	// Has checks if a transaction config exists under the given name
+	Has(name string) (exists bool, err error)
+
+	// Get a transaction config from the store
+	Get(name string) (config transaction.Config, err error)
+
+	// Set stores the given transaction config
+	Set(name string, config transaction.Config) (err error)
+
+	// Delete the transaction config for the given name
+	Delete(name string) (err error)
+
+	// Iter iterates the transaction configs applying the callback for the name and config pairs.
+	// Returning false causes the iteration to stop.
+	Iter(callback func(name string, config transaction.Config) (proceed bool)) (err error)
+}
+
 // Supplier for manager.Store
-type Supplier func(uri string) (s manager.Store, err error)
+type Supplier func(uri string) (s Store, err error)
 
 // Register registers store suppliers
 func Register(name string, s Supplier) (err error) {
@@ -29,7 +52,7 @@ func Register(name string, s Supplier) (err error) {
 
 // New creates a new store with the registered suppliers from the given URI.
 // URI spec: <store>:<arguments>
-func New(uri string) (s manager.Store, err error) {
+func New(uri string) (s Store, err error) {
 	params := strings.SplitN(uri, ":", 2)
 	if len(params) == 0 {
 		return nil, fmt.Errorf("store: invalid uri %s", uri)
